Use abs paths and stdout.  Add docs and tests.
Fix a bug in daemoncmd, where user would specify a relative path to a
pidfile that would causes subsequent commands to fail, b/c the
daemonized process that writes the pidfile had its current working
directory set to '/', placing the pidfile in an unexpected place.  Now
pidfile is coverted to an absolute path before daemonization of the
process.

Previously starting, stopping, and status messages were written to
stderr for no good reason.  Now they are written to stdout, which seems
more conventional, considering they are the output of the command and no
error has occurred.

Tests are low coverage and might not be very robust yet, but something
is better than nothing, right?